+++ src/i18n/languages/csCZ.ts
@@ -38,6 +38,7 @@ const dictionary: RawTranslationPackage = {
     AVERAGE: 'PRŮMĚR',
     AVERAGEA: 'AVERAGEA',
     AVERAGEIF: 'AVERAGEIF',
+    AVERAGEIFS: 'AVERAGEIFS',
     BASE: 'BASE',
     BIN2DEC: 'BIN2DEC',
     BIN2HEX: 'BIN2HEX',