+++ src/i18n/languages/huHU.ts
@@ -38,6 +38,7 @@ const dictionary: RawTranslationPackage = {
     AVERAGE: 'ÁTLAG',
     AVERAGEA: 'ÁTLAGA',
     AVERAGEIF: 'ÁTLAGHA',
+    AVERAGEIFS: 'ÁTLAGHATÖBB',
     BASE: 'ALAP',
     BIN2DEC: 'BIN.DEC',
     BIN2HEX: 'BIN.HEX',