+++ src/i18n/languages/ptPT.ts
@@ -38,6 +38,7 @@ const dictionary: RawTranslationPackage = {
     AVERAGE: 'MÉDIA',
     AVERAGEA: 'MÉDIAA',
     AVERAGEIF: 'MÉDIASE',
+    AVERAGEIFS: 'MÉDIA.SE.S',
     BASE: 'BASE',
     BIN2DEC: 'BINADEC',
     BIN2HEX: 'BINAHEX',